Consolidate common negative code in HttpRequestStream
