+++ src/System.Net.HttpListener/src/System.Net.HttpListener.csproj
@@ -47,6 +47,7 @@
     <Reference Include="System.Threading.Timer" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="System\Net\HttpRequestStream.cs" />
     <Compile Include="System\Net\AuthenticationTypes.cs" />
     <Compile Include="System\Net\ServiceNameStore.cs" />
     <Compile Include="System\Net\HttpListenerRequest.cs" />
